requirements.txt
# 基本ライブラリ
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
requests>=2.26.0

# GUI
tkinter

# Discord Bot
discord.py>=2.0.0

# AI/ML
torch>=1.9.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# OCR関連
nougat-ocr>=0.1.0
easyocr>=1.6.0
pytesseract>=0.3.8

# 音声処理
pydub>=0.25.0
soundfile>=0.10.0

# その他
python-dotenv>=0.19.0
asyncio
threading
queue
subprocess
tempfile
uuid
shutil
